# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Felipe Aragão",
            "email": "felpe.thiago10@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/deals": {
            "post": {
                "description": "Create deal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deals"
                ],
                "summary": "Create deal",
                "parameters": [
                    {
                        "description": "deal request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_deal.DealInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/deals/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a deal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deals"
                ],
                "summary": "Get a deal",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "deal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Deal"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a deal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deals"
                ],
                "summary": "Update a deal",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "deal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "deal request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/update_deal.DealInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_user.UserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_user.UserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}/change-password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change password a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Change password a user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/change_password.UserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "change_password.UserInputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "passwordNew": {
                    "type": "string"
                }
            }
        },
        "create_deal.DealInputDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string"
                        },
                        "city": {
                            "type": "string"
                        },
                        "lat": {
                            "type": "number"
                        },
                        "lng": {
                            "type": "number"
                        },
                        "state": {
                            "type": "string"
                        },
                        "zip_code": {
                            "type": "integer"
                        }
                    }
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trade_for": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "urgency": {
                    "type": "object",
                    "properties": {
                        "limit_date": {
                            "type": "string"
                        },
                        "type": {
                            "type": "integer"
                        }
                    }
                },
                "user_id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "create_user.UserInputDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "location": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string"
                        },
                        "city": {
                            "type": "string"
                        },
                        "lat": {
                            "type": "number"
                        },
                        "lng": {
                            "type": "number"
                        },
                        "state": {
                            "type": "string"
                        },
                        "zip_code": {
                            "type": "integer"
                        }
                    }
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "entity.Deal": {
            "type": "object",
            "required": [
                "description",
                "location",
                "type",
                "urgency",
                "user_id",
                "value"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 3
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/entity.LocationDeal"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trade_for": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "urgency": {
                    "$ref": "#/definitions/entity.Urgency"
                },
                "user_id": {
                    "type": "string"
                },
                "value": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "entity.LocationDeal": {
            "type": "object",
            "required": [
                "address",
                "city",
                "lat",
                "lng",
                "state",
                "zip_code"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "state": {
                    "type": "string",
                    "maxLength": 2,
                    "minLength": 2
                },
                "zip_code": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 10000000
                }
            }
        },
        "entity.LocationUser": {
            "type": "object",
            "required": [
                "address",
                "city",
                "lat",
                "lng",
                "state",
                "zip_code"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "state": {
                    "type": "string",
                    "maxLength": 2,
                    "minLength": 2
                },
                "zip_code": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 10000000
                }
            }
        },
        "entity.Urgency": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "limit_date": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "required": [
                "email",
                "location",
                "login",
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/entity.LocationUser"
                },
                "login": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "handlers.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "update_deal.DealInputDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string"
                        },
                        "city": {
                            "type": "string"
                        },
                        "lat": {
                            "type": "number"
                        },
                        "lng": {
                            "type": "number"
                        },
                        "state": {
                            "type": "string"
                        },
                        "zip_code": {
                            "type": "integer"
                        }
                    }
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trade_for": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "urgency": {
                    "type": "object",
                    "properties": {
                        "limit_date": {
                            "type": "string"
                        },
                        "type": {
                            "type": "integer"
                        }
                    }
                },
                "user_id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}