# internal/infra/repository/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.15.0

package db

import (
	"database/sql"
	"time"
)

type Deal struct {
	ID               string
	UserID           string
	Type             sql.NullInt32
	Value            float64
	Description      sql.NullString
	TradeFor         sql.NullString
	LocationLat      sql.NullFloat64
	LocationLng      sql.NullFloat64
	LocationAddress  string
	LocationCity     string
	LocationState    string
	LocationZipCode  int64
	UrgencyType      int32
	UrgencyLimitDate time.Time
	Photos           sql.NullString
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type User struct {
	ID              string
	Name            string
	Email           string
	Login           string
	Password        string
	LocationLat     sql.NullFloat64
	LocationLng     sql.NullFloat64
	LocationAddress string
	LocationCity    string
	LocationState   string
	LocationZipCode int64
	CreatedAt       time.Time
	UpdatedAt       time.Time
}